Fix address data
